refactor : create and apply accommodation adapter
