+++ src/main/kotlin/com/jaranda/goorlebackend/infra/accommodations/repository/TagRepository.kt
@@ -1,4 +1,4 @@
-package com.jaranda.goorlebackend.domain.accommodations.repository
+package com.jaranda.goorlebackend.infra.accommodations.repository
 
 import com.jaranda.goorlebackend.domain.accommodations.entity.Tag
 import org.springframework.data.jpa.repository.JpaRepository
